requirements.txt
# ACGS-PGP Main Requirements
# Version: 3.0.0
# Last Updated: 2025-06-28
#
# This file is maintained for compatibility.
# Primary dependency management is now handled via pyproject.toml and uv.
# To install: uv pip install -e .

# Core Framework
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.10.5
pydantic-settings>=2.7.1
httpx>=0.28.1

# Database & Storage
redis>=5.0.1
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.7

# Authentication & Security
cryptography>=45.0.4
pyjwt[crypto]>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.10

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-instrumentation-fastapi==0.55b1

# AI/ML Models & Frameworks
anthropic>=0.8.0
openai>=1.3.0
google-generativeai>=0.3.0
groq>=0.4.0
torch>=2.4.0,<2.8.0
transformers>=4.35.0
sentence-transformers>=2.2.2
numpy>=1.24.0,<=2.2.6
scipy>=1.10.0

# NeMo-Skills (install separately with: uv pip install git+https://github.com/NVIDIA/NeMo-Skills.git)
# nemo-skills @ git+https://github.com/NVIDIA/NeMo-Skills.git

# Utilities & CLI
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.6.0
typer>=0.9.0
jinja2>=3.1.0
aiofiles>=23.0.0

# Data Processing
pandas>=2.0.0
polars>=0.20.0
pyarrow>=12.0.0

# Networking & Communication
websockets>=11.0.0
aioredis>=2.0.0
aiohttp>=3.9.0
celery>=5.3.0

# Configuration & Validation
jsonschema>=4.17.0
marshmallow>=3.19.0
cerberus>=1.3.4

# Testing (install with: uv pip install -e .[test])
# pytest>=7.4.3
# pytest-asyncio>=0.21.1
# pytest-cov>=4.1.0
# pytest-mock>=3.14.1
# pytest-xdist>=3.7.0
# pytest-faker>=2.0.0
